% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajusta_beta.R
\name{ajusta_beta}
\alias{ajusta_beta}
\title{Ajuste do Modelo de Regressão Linear}
\usage{
ajusta_beta(formula, data)
}
\arguments{
\item{formula}{Recebe a fórmula que será utilizada no ajuste do modelo.
Deve ser escrita da seguinte maneira:
\verb{Variável Resposta ~ Variáveis Preditoras}.
Exemplo: se \code{Y} é a variável resposta e \code{X1}, \code{X2} e \code{X3} são as variáveis
preditoras, então \code{formula = Y ~ X1 + X2 + X3}.}

\item{data}{Um objeto contendo o banco de dados que será utilizado no ajuste.}
}
\description{
Calcula os valores dos coeficientes (betas) de uma regressão linear com variáveis preditoras numéricas.
}
\examples{
ajusta_beta(Y ~ X1 + X2 + X3, data = banco_exemplo)
}
