% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_pred.R
\name{grafico_pred}
\alias{grafico_pred}
\title{Gráfico de Valores Observados e Preditos}
\usage{
grafico_pred(model, new_data = NULL, file=NULL, width=7, height=7)
}
\arguments{
\item{model}{um objeto de modelo da classe "modelo_linear" cuja predição é desejada.}

\item{new_data}{argumento opcional com conjunto de preditoras para usar no gráfico.}

\item{file}{argumento opcional com o caminho de onde se quer salvar o gráfico.}

\item{width}{argumento opcional com a largura do gráfico em polegadas.}

\item{height}{argumento opcional com a altura do gráfico em polegadas.}
}
\description{
Plota gráfico de valores preditos vs. valores obervados.
}
\examples{
modelo <- ajusta_beta(Y ~ X1 + X2 + X3, data = banco_exemplo)
grafico_pred(modelo)
}
