% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicao.R
\name{predicao}
\alias{predicao}
\title{Predição para Novos Dados}
\usage{
predicao(model, new_data)
}
\arguments{
\item{model}{um objeto da classe "modelo_linear" cuja predição é desejada.}

\item{new_data}{conjunto de valores dos novos preditoras, podendo ser um
data frame ou uma matriz. Deve conter as suas colunas (preditores) nomeadas
com os mesmos nomes das do modelo ajustado. Caso este parâmetro seja mal
especificado, isto é, falta de algum preditor utilizado no ajuste, será
retornado um erro.}
}
\description{
É uma função que faz predições em um modelo linear para um conjunto de dados.
}
\examples{
modelo <- ajusta_beta(Y ~ X1 + X2 + X3, data = banco_exemplo)
predicao(modelo, data.frame(X1 = c(12,15), X2 = c(24,32), X3 = c(8,12)))
}
